src-release: Strip "-cvs" suffix in version number.
For some reason not understood, src-release produces tarballs where
the version number in the tarball name includes the "-cvs" suffix.
Looking at the snapshots produced in the past for CVS snashots,
the "-cvs" suffix was NOT included. This new -cvs suffix is therefore
now causing a mismatch between what the scripts expect and what this
makefile produces. This patch fixes the issue by stripping the "-cvs"
suffix, if present, from the version number.

ChangeLog:

        * src-release (VER): Strip any "-cvs" suffix from the version
        number, if extracted from $(TOOL)/version.in.